Add url param and request body validation for API endpoints (#356)
This resolves #245 by addding API input validation

* Add url param and request body validation for API endpoints
* Add create battle API request body validation
* Sanitize email during login attempt
* Sanitize email on password reset request
* Start adding API tests
* Add api test for create battle
* Apply prettier formatting to e2e tests
* Add retro list and create api tests
* Add storyboard list and create api tests
* Add API tests for org dept team list and create

+++ api/apikey.go
@@ -23,6 +23,11 @@ func (a *api) handleUserAPIKeys() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		UserID := vars["userId"]
+		idErr := validate.Var(UserID, "required,uuid")
+		if idErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, idErr.Error()))
+			return
+		}
 
 		APIKeys, keysErr := a.db.GetUserApiKeys(UserID)
 		if keysErr != nil {
@@ -35,7 +40,7 @@ func (a *api) handleUserAPIKeys() http.HandlerFunc {
 }
 
 type apikeyGenerateRequestBody struct {
-	Name string `json:"name"`
+	Name string `json:"name" validate:"required"`
 }
 
 // handleAPIKeyGenerate handles generating an API key for a user
@@ -54,6 +59,11 @@ func (a *api) handleAPIKeyGenerate() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		UserID := vars["userId"]
+		idErr := validate.Var(UserID, "required,uuid")
+		if idErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, idErr.Error()))
+			return
+		}
 
 		body, bodyErr := ioutil.ReadAll(r.Body)
 		if bodyErr != nil {
@@ -65,6 +75,12 @@ func (a *api) handleAPIKeyGenerate() http.HandlerFunc {
 		jsonErr := json.Unmarshal(body, &k)
 		if jsonErr != nil {
 			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, jsonErr.Error()))
+			return
+		}
+
+		inputErr := validate.Struct(k)
+		if inputErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, inputErr.Error()))
 			return
 		}
 
@@ -90,7 +106,7 @@ func (a *api) handleAPIKeyGenerate() http.HandlerFunc {
 }
 
 type apikeyUpdateRequestBody struct {
-	Active bool `json:"active"`
+	Active bool `json:"active" validate:"required,boolean"`
 }
 
 // handleUserAPIKeyUpdate handles updating a users API key
@@ -110,6 +126,11 @@ func (a *api) handleUserAPIKeyUpdate() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		UserID := vars["userId"]
+		idErr := validate.Var(UserID, "required,uuid")
+		if idErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, idErr.Error()))
+			return
+		}
 		APK := vars["keyID"]
 
 		body, bodyErr := ioutil.ReadAll(r.Body)
@@ -122,6 +143,12 @@ func (a *api) handleUserAPIKeyUpdate() http.HandlerFunc {
 		jsonErr := json.Unmarshal(body, &k)
 		if jsonErr != nil {
 			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, jsonErr.Error()))
+			return
+		}
+
+		inputErr := validate.Struct(k)
+		if inputErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, inputErr.Error()))
 			return
 		}
 
@@ -151,6 +178,11 @@ func (a *api) handleUserAPIKeyDelete() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		UserID := vars["userId"]
+		idErr := validate.Var(UserID, "required,uuid")
+		if idErr != nil {
+			a.Failure(w, r, http.StatusBadRequest, Errorf(EINVALID, idErr.Error()))
+			return
+		}
 		APK := vars["keyID"]
 
 		APIKeys, keysErr := a.db.DeleteUserApiKey(UserID, APK)
